fix linux 32bit int

+++ pandas/tests/series/test_npfuncs.py
@@ -29,7 +29,7 @@ def test_numpy_argwhere(index):
 
     s = Series(range(5), index=index, dtype=np.int64)
 
-    result = np.argwhere(s > 2)
+    result = np.argwhere(s > 2).astype(np.int64)
     expected = np.array([[3], [4]], dtype=np.int64)
 
     tm.assert_numpy_array_equal(result, expected)
